Refactoring and organizing files
