+++ src/js/components/textmarkers/BooleanMarker.jsx
