+++ src/js/components/textmarkers/color/ColorMarker.jsx
