+++ src/js/components/textmarkers/color/color.js
