+++ src/js/components/textmarkers/vector/TrackballControls.js
